Primeiro sistema funcional
Correções de erros e bugs

+++ utilities/func/__init__.py
@@ -17,93 +17,99 @@ def soma(val):
 
 def valor_conta(mesa=-999):
     if mesa == -999:
+        print('-1 para sair')
         mesa = numero_mesa()
-    a = open(str(mesa), 'r')
-    soma = desconto = 0
-    for i in a:
-        i = i.split(';')
-        if i[0].isnumeric():
-            val_total = float(i[2]) * int(i[3])
-            soma += val_total
-        else:
-            desconto += float(i[1])
-    return soma - desconto
+    if mesa != -1:
+        a = open(str(mesa), 'r')
+        soma = desconto = 0
+        for i in a:
+            i = i.split(';')
+            if i[0].isnumeric():
+                val_total = float(i[2]) * int(i[3])
+                soma += val_total
+            else:
+                desconto += float(i[1])
+        return soma - desconto
 
 
 def mostrar_conta(mesa=-999):
     if mesa == -999:
+        print('-1 para sair')
         mesa = numero_mesa()
-    a = open(str(mesa), 'r')
-    itens = []
-    soma = desconto = 0
-    for i in a:
-        itens.append(i)
-    a.close()
-    print(60 * '-')
-    print(f'{"Conta mesa ":>35}{mesa}')
-    print(60 * '-')
-    print(f'{"Quant":6}|{"Produto":^29}|{" Val unit "}|{"Val total":^11}|')
-    for i in itens:
-        i = i.split(';')
-        if i[0].isnumeric():
-            val_total = float(i[2]) * int(i[3])
-            soma += val_total
-            print(f'{i[3]:<4} x  {i[1]:<30}  {textos.moeda(float(i[2]))} {textos.moeda(val_total):>10}')
-        else:
-            desconto += float(i[1])
-    print(60*'-')
-    print(f'{"Valor total":>47}:  {textos.moeda(soma)}')
-    if desconto != 0:
-        print(f'{"Desconto":>47}: -{textos.moeda(desconto)}')
+    if mesa != -1:
+        a = open(str(mesa), 'r')
+        itens = []
+        soma = desconto = 0
+        for i in a:
+            itens.append(i)
+        a.close()
         print(60 * '-')
-        print(f'{"Novo total":>47}:  {textos.moeda(soma - desconto)}')
+        print(f'{"Conta mesa ":>35}{mesa}')
+        print(60 * '-')
+        print(f'{"Quant":6}|{"Produto":^29}|{" Val unit "}|{"Val total":^11}|')
+        for i in itens:
+            i = i.split(';')
+            if i[0].isnumeric():
+                val_total = float(i[2]) * int(i[3])
+                soma += val_total
+                print(f'{i[3]:<4} x  {i[1]:<30}  {textos.moeda(float(i[2]))} {textos.moeda(val_total):>10}')
+            else:
+                desconto += float(i[1])
+        print(60*'-')
+        print(f'{"Valor total":>47}:  {textos.moeda(soma)}')
+        if desconto != 0:
+            print(f'{"Desconto":>47}: -{textos.moeda(desconto)}')
+            print(60 * '-')
+            print(f'{"Novo total":>47}:  {textos.moeda(soma - desconto)}')
 
 
 def fechar_conta(mesa=-999):
     if mesa == -999:
+        print('-1 para sair')
         mesa = numero_mesa()
-    while True:
-        if existe_arquivo(str(mesa)):
-            mostrar_conta(mesa)
-            val_conta = valor_conta(mesa)
-            while True:
-                try:
-                    val_pago = float(input('Valor a ser pago  [0: valor total]\n'
-                                                            f'{"[-1: sair]":>28}    :R$'))
-                except:
-                    textos.mensagem_erro('Dígite um valor válido')
-                else:
-                    if val_pago > val_conta:
-                        textos.mensagem_erro('Valor digitado maior que o valor da conta!')
-                    elif val_pago == val_conta or val_pago == 0:
-                        pagamento(val_conta)
-                        deletar_fechar_mesa(mesa)
-                        print('Pagamento realizado com sucesso!')
-                        sair = True
-                        break
-                    elif val_pago > 0 and val_pago < val_conta:
-                        pagamento(val_pago)
-                        a = open(str(mesa), 'a')
-                        a.write(f'{"desconto"};{val_pago};\n')
-                        a.close()
-                        print('Pagamento realizado com sucesso!')
-                        resp = ''
-                        while True:
-                            resp = input('Adicionar mais um pagamento a mesa? (S/N):').lower()
-                            sair = False
-                            if resp == 'n':
-                                sair = True
-                                break
-                            elif resp == 's':
-                                break
-                            else:
-                                textos.mensagem_erro(' Valor inválido!')
-                        break
-            if sair:
-                break
-        else:
-            textos.mensagem_erro(' Mesa fechada!')
-            mesa = numero_mesa()
+    if mesa != -1:
+        while True:
+            if existe_arquivo(str(mesa)):
+                mostrar_conta(mesa)
+                val_conta = valor_conta(mesa)
+                while True:
+                    try:
+                        val_pago = float(input('Valor a ser pago  [0: valor total]\n'
+                                                                f'{"[-1: sair]":>28}    :R$'))
+                    except:
+                        textos.mensagem_erro('Dígite um valor válido')
+                    else:
+                        if val_pago > val_conta:
+                            textos.mensagem_erro('Valor digitado maior que o valor da conta!')
+                        elif val_pago == val_conta or val_pago == 0:
+                            pagamento(val_conta)
+                            deletar_fechar_mesa(mesa)
+                            print('Pagamento realizado com sucesso!')
+                            sair = True
+                            break
+                        elif val_pago > 0 and val_pago < val_conta:
+                            pagamento(val_pago)
+                            a = open(str(mesa), 'a')
+                            a.write(f'{"desconto"};{val_pago};\n')
+                            a.close()
+                            print('Pagamento realizado com sucesso!')
+                            resp = ''
+                            while True:
+                                resp = input('Adicionar mais um pagamento a mesa? (S/N):').lower()
+                                sair = False
+                                if resp == 'n':
+                                    sair = True
+                                    break
+                                elif resp == 's':
+                                    break
+                                else:
+                                    textos.mensagem_erro(' Valor inválido!')
+                            break
+                if sair:
+                    break
+            else:
+                textos.mensagem_erro(' Mesa fechada!')
+                mesa = numero_mesa()
 
 
 def pagamento(valor):
@@ -111,18 +117,20 @@ def pagamento(valor):
         try:
             forma = int(input(f'Digite a forma de pagamento: [1 para dinheiro]\n'
                               f'{"[2 para crédito]":>45}\n'
-                              f'{"[3 para débito]":>44}: '))
+                              f'{"[3 para débito]":>44}\n'
+                              f'{"[-1 para sair]":>43}: '))
         except:
             textos.mensagem_erro('Valor inválido!')
         else:
-            lista = [1,2,3]
-            if forma in lista:
-                a = open('faturamento', 'a')
-                a.write(f'{forma};{valor};\n')
-                a.close()
-                break
-            else:
-                textos.mensagem_erro('Opção inválida!')
+            if forma != -1:
+                lista = [1,2,3]
+                if forma in lista:
+                    a = open('faturamento', 'a')
+                    a.write(f'{forma};{valor};\n')
+                    a.close()
+                    break
+                else:
+                    textos.mensagem_erro('Opção inválida!')
 
 
 def mostrar_faturamento():
@@ -161,27 +169,29 @@ def mostrar_faturamento():
 
 def deletar_mesa(mesa=-999):
     if mesa == -999:
+        print('-1 para sair')
         mesa = numero_mesa()
-    try:
-        os.unlink(str(mesa))
-    except:
-        textos.mensagem_erro(f' Mesa {mesa} não está aberta!')
-    else:
-        a = open('mesas_abertas', 'r')
-        linhas = a.readline().split(';')
-        a.close()
-        b = open('mesas_abertas', 'w')
-        cont = False
-        for i in linhas:
-            if i != str(mesa) and i != '':
-                b.write(f'{i};')
-            else:
-                cont = True
-        b.close()
-        if cont:
-            print(f'Mesa {mesa} exlcuida!')
+    if mesa != -1:
+        try:
+            os.unlink(str(mesa))
+        except:
+            textos.mensagem_erro(f' Mesa {mesa} não está aberta!')
         else:
-            print(f'Mesa {mesa} não está aberta!')
+            a = open('mesas_abertas', 'r')
+            linhas = a.readline().split(';')
+            a.close()
+            b = open('mesas_abertas', 'w')
+            cont = False
+            for i in linhas:
+                if i != str(mesa) and i != '':
+                    b.write(f'{i};')
+                else:
+                    cont = True
+            b.close()
+            if cont:
+                print(f'Mesa {mesa} exlcuida!')
+            else:
+                print(f'Mesa {mesa} não está aberta!')
 
 
 def deletar_fechar_mesa(mesa=-999):
@@ -215,53 +225,61 @@ def numero_mesa():
 
 def abrir_mesa(mesa=-999):
     if mesa == -999:
+        print('-1 para sair')
         mesa = numero_mesa()
-    if existe_arquivo(str(mesa)) is False:
-        a = open('mesas_abertas', 'a')
-        a.write(f'{mesa};')
-        a.close()
-        b = open(str(mesa), 'a')
-        b.close()
-        print(f'Mesa {mesa} aberta com sucesso!')
-    else:
-        print('Mesa já aberta!')
-        sleep(2)
+    if mesa != -1:
+        while True:
+            if existe_arquivo(str(mesa)) is False:
+                a = open('mesas_abertas', 'a')
+                a.write(f'{mesa};')
+                a.close()
+                b = open(str(mesa), 'a')
+                b.close()
+                print(f'Mesa {mesa} aberta com sucesso!')
+                break
+            else:
+                print('\033[31mMesa já aberta!\033[m')
+                sleep(2)
+                print('-1 para sair')
+                mesa = numero_mesa()
 
 
 def add_item_mesa():
+    print('-1 para sair')
     n_mesa = numero_mesa()
-    if existe_arquivo(str(n_mesa)) is False:
-        abrir_mesa(mesa=n_mesa)
-    while True:
-        try:
-            cod = int(input('código do produto:'))
-        except:
-            textos.mensagem_erro('Digite um código válido!')
-        else:
-            existe = False
-            for i in cardapio:
-                if cod == int(i[0]):
-                    existe = True
-                    break
-            if existe:
-                break
+    if n_mesa != -1:
+        if existe_arquivo(str(n_mesa)) is False:
+            abrir_mesa(mesa=n_mesa)
+        while True:
+            try:
+                cod = int(input('código do produto:'))
+            except:
+                textos.mensagem_erro('Digite um código válido!')
             else:
-                textos.mensagem_erro('Produto não encontrado')
+                existe = False
+                for i in cardapio:
+                    if cod == int(i[0]):
+                        existe = True
+                        break
+                if existe:
+                    break
+                else:
+                    textos.mensagem_erro('Produto não encontrado')
 
-    while True:
-        try:
-            quant = int(input('Quantidade:'))
-        except:
-            textos.mensagem_erro('Digite um número válido!')
-        else:
-            break
+        while True:
+            try:
+                quant = int(input('Quantidade:'))
+            except:
+                textos.mensagem_erro('Digite um número válido!')
+            else:
+                break
 
-    for i in cardapio:
-        if cod == int(i[0]):
-            a = open(str(n_mesa), 'a')
-            a.write(f'{i[0]};{i[1]};{i[2]};{quant};\n')
-            a.close()
-            print(f'Produto adicionado com sucesso na mesa {n_mesa}.')
+        for i in cardapio:
+            if cod == int(i[0]):
+                a = open(str(n_mesa), 'a')
+                a.write(f'{i[0]};{i[1]};{i[2]};{quant};\n')
+                a.close()
+                print(f'Produto adicionado com sucesso na mesa {n_mesa}.')
 
 
 def mostrar_mesas():
@@ -285,6 +303,16 @@ def mostrar_mesas():
     print('\033[m')
 
 
+def mostrar_cardapio():
+    textos.titulo('Cardápio')
+    a = open('cardapio', 'r')
+    print(f'| {"Código":^6} | {"Nome":^37} | {"Preço":^7} |')
+    for i in a:
+        linha = i.split(';')
+        print(f'{linha[0]:^10} {linha[1]:<40}{textos.moeda(float(linha[2])):<10}')
+    a.close()
+
+
 def cardapio_add():
     # defining a list to keep the data before add to the archive
     lista = [0, 0, 0]
@@ -292,10 +320,13 @@ def cardapio_add():
     # the three while loops below get the info's needed to register the product
     while True:
         try:
+            print('-1 para sair')
             lista[0] = int(input('Codigo do produto:'))
         except:
             textos.mensagem_erro('Digite um código válido!')
         else:
+            if lista[0] == -1:
+                break
             sair = True
             a = open('cardapio', 'r')
             for i in a:
@@ -306,27 +337,28 @@ def cardapio_add():
             a.close()
             if sair:
                 break
-    while True:
-        try:
-            lista[1] = str(input('Nome:')).capitalize()
-        except KeyboardInterrupt:
-            textos.mensagem_erro('Dígite um nome válido!')
-        else:
-            if ';' in lista[1]:
-                textos.mensagem_erro('Caractere ";" inválido!')
-            break
-    while True:
-        try:
-            lista[2] = float((input('Preço: R$')).replace(',', '.'))
-        except:
-            textos.mensagem_erro('Digite um preço válido')
-        else:
-            break
-    # opening a file to keep the data
-    a = open('cardapio', 'a')
-    a.write(f'{lista[0]};{lista[1]};{lista[2]};\n')  # data are added with a ';' to separate
-    a.close()
-    print('Produto cadastrado com sucesso!')
+    if lista[0] != -1:
+        while True:
+            try:
+                lista[1] = str(input('Nome:')).capitalize()
+            except KeyboardInterrupt:
+                textos.mensagem_erro('Dígite um nome válido!')
+            else:
+                if ';' in lista[1]:
+                    textos.mensagem_erro('Caractere ";" inválido!')
+                break
+        while True:
+            try:
+                lista[2] = float((input('Preço: R$')).replace(',', '.'))
+            except:
+                textos.mensagem_erro('Digite um preço válido')
+            else:
+                break
+        # opening a file to keep the data
+        a = open('cardapio', 'a')
+        a.write(f'{lista[0]};{lista[1]};{lista[2]};\n')  # data are added with a ';' to separate
+        a.close()
+        print('Produto cadastrado com sucesso!')
 
 
 def cardapio_del():
